restapi: V3 compatibility support for scheduling policies
In version 4 of the API the support for specifying the scheduling policy
name and parameters as part of the cluster was removed. Instead of that
now it is necessary to specify the identifier of a previously created
scheduling policy. Now that the API supports multiple versions this
support can be rea-added without modifying the V4 implementation. If the
identifier of the scheduling policy isn't provided, the server will
automatically select a scheduling policy that is compatible with the
parameters given by the caller. If no such scheduling policy exists an
error response will be generated.

Change-Id: Ied469fb1b00130c5a2ea21409d3d4157bb844a73
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>