core: Validate steps using CoCo on RunVmCommand
When RunVmCommand calls CreateAllsnapshotsFromVmCommand with
execution context it should use getContext instead of cloneContext
so the executionContext of RunVmcommand should also get updated with the
relevant steps.

The reason why this should be changed is because CoCo is using the
executionContext of the command itself while the AsyncTaskManager sets
the executionContext of the child, CreateAllsnapshotsFromVmCommand, once
RunVmCommand is being called.

Once the engine also sets the context of the existing command, RunVmCommand,
the engine will not throw an NPE that the step does not exists.

Change-Id: I6dfed690d8a15684beda506808e788a2be805aa6
Bug-Url: https://bugzilla.redhat.com/1305809
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>