core: RunVmCommand using callbacks
Set the commnad id and the shouldBeEndedByParent with false.

RunVmCommand is calling its only child command,
CreateAllSnapshotsFromVmCommand, form the execute command
and at the endAction phase.
Once the command is used with a callback we need to change the command
to be set with ShouldBeEndedByParent as false so the snapshot and the
disk will be unlocked.

Change-Id: I6a5889eb4ab378347baa46672afaf2edf48cd738
Bug-Url: https://bugzilla.redhat.com/1305809
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>