core: VmDeviceMonitoring save hash after updateFromFullList()
In VmDeviceMonitoring, after updateFromFullList(), on the next polling
cycle only devices hash is updated, FullList is not called again.
But, because isVmDevicesChanged() simply returned false and the VM
wasn't touched, the new hash wasn't stored in the DB.

This patch adds new return value for isVmDevicesChanged() - HASH_ONLY -
to update hash in the DB without refreshing the device list. For the
same goal added a new list of VMs whose hash needs to be saved.

Change-Id: I5daad3d73f67a578cd0df4b56b104cf791a5fe2e
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>