core: LiveMerge - avoid unneeded endAction() run
When running the live merge flow, the endAction() of
RemoveSnapshotSingleDiskLiveCommand shouldn't run as currently
the logic is contained within the onFailed()/onSucceeded() methods
(that should be changed in a separate change).

Furthermore, in this change the end methods are overriden to clarify that
the logic in the parent command shouldn't be executed ever.

Change-Id: I562d07d713cba53c0ca1a2c2a14f05d65d98ab8d
Bug-Url: https://bugzilla.redhat.com/1306743
Signed-off-by: Liron Aravot <laravot@redhat.com>