core: Move getImageInfo call outside of transaction boundary
Currently, when syncing db records, we call vdsm to get volume info.
This call is performed inside db transaction which could lead to failure
regardless of the db performance. This could happen if storage is slow or
network connection is slow.
To fix this behavior, the call to vdsm moved outside of the transaction.

This patch will simplify fixing the related bug.

Change-Id: Ifac1ab925afd448d149fecc286d361a797c9d57b
Bug-Url: https://bugzilla.redhat.com/1309294
Signed-off-by: Ala Hino <ahino@redhat.com>