engine: Fix migration policy update for running VM
Problem description:
--------------------
When a VM is running and its migration policy is changed to pinned - OR
A pinned VM is running and it dedicated hosts list is modified.
The VM may running on un-dedicated host, locking the VM away from its
dedicated host. Requiring VM reboot.

Desired behavior:
-----------------
1. When a VM is running and changing migration policy from un-pinned
  to pinned.
  - Validate the VM is running on one of the targeted pinned hosts,
  - Fail with adequate warning.
  - Proceed with normal update.
2. When pinned VM is running and changing the list of dedicated hosts.
  And the dedicated hosts list excluding the pinned host.
  - Validate the VM is running on one of the targeted pinned hosts,
  - Fail with adequate warning.
  - Proceed with normal update.

Implementation:
--------------
Add canDoAction constraint on modify a pinned VM, if it is running -
must run on its dedicated host. (UpdateVmCommand.java)

Change-Id: Ib3dc32a2b92ea0d111328eb6a6803fa91c015836
Bug-Url: https://bugzilla.redhat.com/1058384
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>