restapi: Preserve parameters when forwarding
The version filter forwards requests to specific versions of the API,
adding the "/v3" or "/v4" prefix to the URL as needed. But when it does
this it also loses the matrix and query parameters. For example, if the
original URI is this:

  /hosts/{host:id};force=true

The forwarded URI is this, with the version prefix but without the
parameters:

  /v4/hosts/{hosts:id}

This patch fixes that error.

Change-Id: I4706634c26c3ea60c73cef5f2d070344e488b349
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>