core: simplify the creation of vm-analyzer's factory
The factory is created by VmsMonitoring on each monitoring cycle,
including cycles that were triggered by events. On large scale,
the frequency can be high. Therefore, this patch removes the
injections to VmAnalyzerFactory and instead the dependencies
that are injected only once to the VmsMonitoring are passed to
its constructor.

Change-Id: I9a55c055212cac958da0fdbf86ef77bac60c20a8
Signed-off-by: Arik Hadas <ahadas@redhat.com>