core: storage_domain_shared_status SQL calculation
A storage domain's shared status was calculated by the
fn_get_storage_domain_shared_status_by_domain_id function, which posed
several performance issues, as it forced the switching between pgplsql
and sql contexts several times, not to mention performing DML and even
potentially DDL operations as part of the flow.

Additionally, this function was phrased in a convoluted and hard to
understand fashion, making it hard to maintain. The old logic, before
this patch, was as follows:

IF (domain.type != ISO) THEN
        IF (domain not attached to a pool) THEN
                RETURN Unattached
        ELSE IF (domain.status == Active) THEN
                RETURN Active
        ELSE
                RETURN Inactive
        ENDIF
ELSE
        IF (num of distinct statuses == 0) THEN
                RETURN Unattached
        ELSE IF (num of distinct statuses == 1) THEN
                IF (domain.status == Active) THEN
                        RETURN Active
                ELSE
                        RETURN Inactive
                ENDIF
        ELSE
                IF (domain is active in at lest one pool) THEN
                        RETURN Mixed
                ELSE
                        RETURN Inactive
                END
        END
END

This logic can clearly be simplified. First and foremost, the special
treatment for an ISO domain should be removed. Instead, a general
treatment for a domain with multiple pools should be set in place, where
ISO just happens to be the only type that supports this. The refactored
logic should handle two variables: the number of distinct statuses a
domain has and the number of active statuses it has:

IF (num_distinct_statuses) == 0 THEN
        RETURN Unattached
ELSE IF (num_active_statuses > 0) THEN
        IF (num_distinct_statuses > 1) THEN
                RETURN Mixed
        ELSE
                RETURN Active
ELSE
        RETURN Inactive
END

This logic can easily be expressed as an SQL query which can be joined
on and save the context switches associated with a pgplsql function.

Change-Id: I9c1695a41e7a9ca3fb32cbbbfc8fc8e4ae80f6ef
Signed-off-by: Allon Mureinik <amureini@redhat.com>