core: extract balloon related code from vm-analyzer
VmAnalyzer should (ideally) take a VM that was reported by VDSM,
compare it to the previous information we have for that VM and
analyze the difference. However, the analyzer currently does
more than that and therefore is not easy to maintain.

In this patch things that are related to monitoring balloon
information are extracted from VmAnalyzer to a new class called
BalloonMonitoring. The benefits are:
1. Less not-that-related code in VmAnalyzer and one less DAO
to pass to it.

2. No use of static fields in VmAnalyzer (VmsMonitoring is
already a singleton).

3. We get to the point where the dynamic data of the VM is
persisted to the database faster - we eliminate a call to the
DAO per-VM before the data is flushed into the database. This
is now done after the important information is saved into the
database.

Change-Id: I3ccdab942209ea83f7803f4736cc49f371aa23dc
Signed-off-by: Arik Hadas <ahadas@redhat.com>