core: extract luns related code from vm-analyzer
VmAnalyzer should (ideally) take a VM that was reported by VDSM,
compare it to the previous information we have for that VM and
analyze the difference. However, the analyzer currently does
more than that and therefore is not easy to maintain.

This patch extracts the LUN disks handling from VmAnalyzer to
a new class called LunDisksMonitoring. The benefits are:
1. Less not-that-related code in VmAnalyzer and one less DAO to
pass to it.

2. We get to the point where the up-to-date dynamic data of the
VM is persisted to the database faster as we push one database
call per VM that is powering up to be the dynamic VM data is
save to the database.

Change-Id: I6a5f0410a0684c8785c08d27ed88502822158e85
Signed-off-by: Arik Hadas <ahadas@redhat.com>