core: extract vm jobs related code from vm-analyzer
VmAnalyzer should (ideally) take a VM that was reported by VDSM,
compare it to the previous information we have for that VM and
analyze the difference. However, the analyzer currently does
more than that and therefore is not easy to maintain.

This patch extracts the VM jobs handling from VmAnalyzer to a
new class called VmJobsMonitoring. The benefits are:
1. Less not-that-related code in VmAnalyzer and one less DAO
to pass to it.

2. We get to the point where the dynamic VM data is persisted
into the database faster as we push the queries that are done
in order to process the reported VM jobs to be after the VM
dynamic data is persisted to the database.

3. The processing of VM jobs is done by another module which
is a singleton and therefore it prepares the ground for
events-based updates of command callbacks and for possible
optimizations of the inefficient queries from the database
that are done in statistics monitoring cycles.

Change-Id: Ic244bfc0ca2aef014b09aaa2fcf4a1d7f1e9d166
Signed-off-by: Arik Hadas <ahadas@redhat.com>