frontend: Only warn on timezone difference for Windows VMs
Due to the fact that Linux systems are recommended to use UTC as timezone
for the RTC clock, the check is only useful for Windows VMs due to the fact
that normally the local timezone is stored in the RTC clock for windows.

The warning basically will indicate a discrepancy between what's configured
in the VM dialog as RTC time offset and what's actually configured in the OS.
Since the value will be reset to what's set in the VM Dialog that is a good
thing to warn the users about.

Change-Id: I5b77a412f2a5b3929e7a8dface0e9273fdec8fdb
Bug-Url: https://bugzilla.redhat.com/1281871
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>