webadmin: Populate host protocol on 'edit host'
In cluster level >= 3.6, only jsonrpc protocol is supported.
In cluster level 3.5 both xmlrpc and jsonrpc are supported.
In cluster level <= 3.4 only xmlrpc is supported.

The UI should be able to set 'jsonrpc' as a protocol
for the host when the cluster is set to 3.6. To allow the user
to select the desired protocol in 3.5 and set protocol to
xmlrpc in 3.4.

Change-Id: I51c50e7311ef536785cfc06f5e640fec2262f6f3
Bug-Url: https://bugzilla.redhat.com/1289868
Signed-off-by: Moti Asayag <masayag@redhat.com>