core: Remove condition for AsyncTaskType.notSupported.
In some cases the engine calls commands which
have async calls to VDSM but in some scenarios,
those async calls, eventually are not getting called.

In case that the async calls are not getting called,
the engine should set the command as succeeded since
there are no tasks it should wait for to determine
the command's status.

Change-Id: I49b7f506dc1142c3a559d0e6f64813e2ade02254
Bug-Url: https://bugzilla.redhat.com/1310390
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>