webadmin: Do not fetch disk info together with VM templates
In GetVmTemplatesByStoragePoolId and GetVmTemplatesFromStorageDomain
queries, the disk info fetched by them is not used in most situations.
However, fetching the excessive data takes time and slows down the UI.
This patch gets rid of disk info where it is not needed.

In particular, in 'New VM' and similar dialogs the disk info for the
currently selected template is queried separately, while the disk info
fetched previously for every template in the list is not used at all.

Change-Id: I6fa007c1cc7b49efb3d9d4e1c01a852aa8a3cf62
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1308478
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>