core: LiveMerge: Remove audit message when delete succeeds
Currently, if live merge succeeds we get an audit message
indicating that volume doesn't exist. This hapens during
DESTROY_IMAGE_CHECK step where we verify that image removed. In a
successful use case, the image is indeed removed hence, we get an
audit message alerting that volume doesn't exist.  To fix this, we
added an option to indicate whether we want to log audit message.

Change-Id: I18d190aa05c7f1e404c10b11fd537e92ba77e434
Bug-Url: https://bugzilla.redhat.com/1310615
Signed-off-by: Ala Hino <ahino@redhat.com>