restapi: Don't fail with unexpected version numbers
Currently the API assumes that the format of version numbers is
something like 4.3.1-100, and it fails due to an array index out of
bounds exception if it has a different format. This patch makes this
code more tolerant, and avoids the failures.

Change-Id: Icaf19907ee09b8e4db731c3516520abce2527d50
Bug-Url: https://bugzilla.redhat.com/1311616
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 430b799dc656e51244ae97a32340370f4469c6ad)