core: packaging: Remove dependency on org.tukaani.xz
Removes dependency on org.tukaani.xz which is optional dependency of
apache-common-compress to support XZ compression. We don't use XZ
compression inside engine directly and apache-commmons-compress RPM
is part of EL7, so we no longer need to include it inside oVirt/RHEV
RPMs, so we no longer need to have this optional dependency to
satisfy legal requirements.

Change-Id: Icb86e2e03d2aa468dbac1732b9944658c7cdd027
Signed-off-by: Martin Perina <mperina@redhat.com>