core: CommandBase - avoid setting end status prematurely
In change I49b7f506dc1142c3a559d0e6f64813e2ade02254 the code checking whether
the command should status should be updated after execution was updated.

As some of the commands are adding the tasks to the internal task id list
instead of using the getTaskIdList() method it'll cause those commands to be marked
as ended although the tasks are still running.

This patch adds a check to the internal list as well to avoid that issue.

Change-Id: I2251dd808734f297c5e3907b9f106d6cf87fbd18
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1310390
Signed-off-by: Liron Aravot <laravot@redhat.com>