restapi: Ensure that V4 delegates are always non-null
The V4 compatibility layer is composed by a set of server classes that
are almost empty, as they forward almost all the requests to the
corresponding V4 "delegates". The delegate is passed to the V4
constructor, and it is assumed to be always be non-null, but that isn't
explicitly checked. Some code coverage tools complain about the fact
that it can in theory be null. This patch adds an assertion to the
constructors, which will hoefully stop this complaining.

Change-Id: I7b1405a4856c9b4344f738864b3f18dfd5660da8
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>