core: early return in batch operations with no elements
Seems like currently on every batch operation (save/update/remove) we use a
connection to the database. Not all the callers to these operations verify
that there are actually elements in the collection of entities they pass.
Therefore, we go deep into the stack and get a connection to the database
in order to execute stored procedure that doesn't do anything.

As a solution, this patch suggests to return early without doing anything
in these operations when the passed collection of entities is empty.

Change-Id: I31e10d048e9f6bf63bbc2ac51e330633c7dc3423
Signed-off-by: Arik Hadas <ahadas@redhat.com>