core: prevent duplicate virtio-serial device
In 3.6 the virtio-serial device becomes managed device. However, it
can still be reported as unmanaged if the VM was started in pre 3.6
version and was running during the upgrade and its ID will be
different than the one stored in the database.

This patch suggests to look for virtio-serial devices not only by
their ID but by their type as well. That way, when we get unamanged
virtio-serial reported by VDSM, we will be able to correlate it to
the managed device and it will prevent adding it again.

This solution assumes that only one virtio-serial device is defined
for each VM.

Change-Id: I3546332024b6ee7667340f031c6c225e1ee67ead
Bug-Url: https://bugzilla.redhat.com/1308885
Signed-off-by: Arik Hadas <ahadas@redhat.com>