webadmin: edit running VM is not possible
It was not possible to edit the running VM. The reason was that the
VmNextRunConfigurationModel in constructor always inited the latch which by
default had the 'available' flag set to true (even not shown since this flag
was set sooner than the listener which would actually show it was registered).

As a consequence, the validate() method returned false, since the latch was
"available" and unchecked. But since it was not rendered, it was not obvious to
the user what the problem is making the dialog impossible to submit.

The fix was just to move the initialization of the latch to the proper place.

Change-Id: I13f9b6d072c2ba8a2113936e95223df9b9f4db7e
Bug-Url: https://bugzilla.redhat.com/1312636
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>