core: Clean ActionVersionMap's DAO
ActionVersionMap's are not (and should not!) be update from within the
application, only by upgrade scripts. Hence, there's no reason to have a
DAO that can manipulate them, and these methods are indeed not used
outside of the test's scope.

This patch removes these redundant methods, as well as the stored
procedures that back them up.

Change-Id: Iff6f8ea4a8ee81c60d983a6973bb3334d3e883a0
Signed-off-by: Allon Mureinik <amureini@redhat.com>