core: Provisioning of VMs created from Templates tab fix
Previously when VM was create from Templates tab ("New VM"), its type
was Desktop and "Optimized for" listbox was not touched in "New Virtual
Machine" dialog then the VM was created with disk provisioning "Clone"
which corresponds to vm type "Server". It could be workaround by
selecting "Server" and "Desktop" option back again in "Optimized for"
listbox.

This patch fixes that so VMs optimized for Desktop created from
Templates tab withoud touching "Optimized for" are correctly
provisioned - Thin.

Bug-Url: https://bugzilla.redhat.com/1309221
Change-Id: I7d3a3b3afbc933135121f6bd02ca3fe665edfdad
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>