restapi: Ensure that V4 delegates are always non-null
The V4 compatibility layer is composed by a set of server classes that
are almost empty, as they forward almost all the requests to the
corresponding V4 "delegates". The delegate is passed to the V4
constructor, and it is assumed to be always be non-null, but that
isn't explicitly checked. Some code coverage tools complain about the fact
that it can in theory be null. This patch changes the constructor that
receives this delegate so that it throws an exception if it is null. It
also changes the visibility of the field, making it private, so it is
impossible to change its value after the object is created. These to
measures should make it clear for any code coverage tool that it is
impossible for this field to be true.

Change-Id: I1060f8258443db8622ba6069e3659c71c8179ed5
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>