restapi: Check dispatcher in version filter
Currently the version filter doesn't check the result of the
"getRequestDispatcher" method, which may be null when there is no such
dispatcher. If this happens that will generate a NPE. That isn't a
serious problem, as it will result in an HTTP 500 (internal server error)
code sent to the client for a request that is incorrect anyhow. But it
is better to check it, write a message in the log, and send to the
client a HTTP 404 (not found) code.

Change-Id: Ib73a6ef557f079f667bbddca6596799a08c5ddad
Coverity-Id: 1352163
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>