restapi: Tranform correctly arrays of bytes
Some of the methods of the system resource return arrays of bytes
instead of objects. For example, the "?schema" requests returns the XML
schema as an array of bytes. Currently this produces an exception
because the V3 output adapter doesn't know how to handle arrays of
bytes. This patch fixes that issue.

Change-Id: Id9476bd0ac0da315af27d9f63fa41bbb9fe1b7dd
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>