core: manage vm-jobs in-memory
Previously the process was:
MergeCommand -> persist job into the database
Monitoring -> on each cycle read jobs from the database and
compare them with the jobs reported by vdsm

The patch suggests another way:
MergeCommand -> pass the job to the monitoring that persists
it into the database and store it in-memory
Monitoring -> compare the reported jobs with the the jobs
that are stored in-memory

On startup, the monitoring retrieve all the existing job in
the database.

This way, we reduce the number of reads from the database.

Change-Id: Id8a9e935f49f458388c089398b110a8854c197e5
Signed-off-by: Arik Hadas <ahadas@redhat.com>