engine: Remove hotplugged memory & CPUs from changed fields list
If number of CPUs or memory size is changed when VM is running, special
info is displayed in the confirmation dialog about hotplugging CPUs or
memory. In the same time, numOfSockets and memSizeMb fields are present
in the list of fields to be changed on the next VM run. Noting these
fields there is redundant, so it was eliminated.

Note that memSizeMb is removed only in the case of memory plugging, but
not in the case of unplugging. Confirmation dialog currently assumes
that memory hot unplugging is not supported and decreased memory size is
applied only on the next VM run.

Change-Id: I38a52f7bfe4f4b6ab35b07e7526746f748c5b92d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1314426
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>