webadmin: Add External Network Provider to list of providers
External Network provider is a new type of external provider
is oVirt. The UI must be modified to accomodate it, and an
appropriate model must be added.

Several classes in this patch are still empty (like
ExternalNetworkProviderProperties), they will be filled with
content when subsequent features for external networks will
be added.

Change-Id: I66e6898b273afab2333af950e186e2d970e89875
Signed-off-by: mirecki <mmirecki@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1314375