core: Fix collection type in CreateUserSessionParameters
Unfortunately ExtMap type defined in extensions API cannot be easily
included into our GWT build infrastructure, so groups, that user is a
member of (represented as Collection<ExtMap>), are incorrectly defined
as Collection<Map> in CreateUserSessionParameters. This patch corrects
the definition to Collection<? extends Map> as ExtMap is indirect
ancestor of Map.

Change-Id: I935d4aa7867c11335cdaf1bfbe95e3c10f160bf4
Signed-off-by: Martin Perina <mperina@redhat.com>