core: Remove InitStorageSparseSizeInGB ConfigValue
The InitStorageSparseSizeInGB ConfigValue is pointless, as the size
property of the disk already takes this into account. It has been
removed from production code throughout the codebase (although,
surprisingly, it was not removed from the tests), and only appears
mistakenly in the quote calculation of creating a new pool, where it has
no effect anyway, as its value is 1 and it's used in multiplication.

This patch finishes the job and removes all trace of this redundant
config value

Change-Id: I4358a78a88898fc93b095becc883dec5d066246e
Signed-off-by: Allon Mureinik <amureini@redhat.com>