engine: reuse of existingAttachments data memeber
In the begining of the 'HostSetupNetwork.validate()' there
is completion of the 'existingAttachments' data member
and the 'parameters' of the command.
Therefore, if two attachments are the same (existing and
the one in the parameters), the comparison will be able to
detact it. However, if when doing the comparison new version of the
existing attachments is used (fetch new version from the db), the
comparison will fails, since the exisiting attachments may have missing
data (that the completaion should fix).

Change-Id: I33cbf93c00c0e0dd9bab9c218b6472a4a4723c11
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>