restapi: Unified parameter support
Currently request parameters are handled in different ways. Some are
query parameters, some are matrix parameters, and some are headers. This
causes confussion and forces clients (like the SDKs) to behave
differently for different parameters. To avoid this confussion this
patch changes the server so that all parameters work in the same way:
they will be extracted from the query. If they aren't in the query then
they will be extracted from the matrix parameters. Finally, some special
parameters, like "filter", will also be extracted from the HTTP headers,
for backwards compatibility.

Change-Id: I2d39990e0b3c5d8c46dc30a876b60407e7ed343e
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>