webadmin: Fix assigning permission to group
Patch I5b1750301a48c874f979f91ce7680b610ba03516 introduces split views
of users or groups when assigning permissions. Unfortunately current
GWT representation of users and groups is very broken as we use DbUser
class to represent both users and groups. That led to unfortunate
decission to map group's Display Name to loginName attribute and this
mapping broke method isGroup() which uses loginName to distinguish
between user or group (another very broken design), which led to an
error during assigning permissions to a group. This flow didn't insert
group into ad_groups table, which means that this permission is not
loaded on user login.

This patch fixes assignment of a permission to a group by using DbUser
attribute note (which is not used otherwise) instead of loginName (it's
a hack, but at least it doesn't break existing functionality).

Change-Id: I12bc5b0d158d2572883ef8baa65f0aca856ed0a8
Signed-off-by: Martin Perina <mperina@redhat.com>