core: Reindex vdc_options after upgrade
Once installed in a running system, the vdc_options table is
relatively static, with users rarely changing it. However, upgrading
the system usually involves a large amount of changes to it (see
details below). This makes this table a good candidate for indexing at
the end of the upgrade (hence the script is added to the post_upgrade
directory).

Consider, e.g., upgrading from the latest 3.6 release (i.e., 3.6.3.4,
built from commit cb35e8c) to the current master branch (commit
7fe2363, which will eventually be a part of 4.0 when it's released):
In 3.6.3, this table had 871 rows. The upgrade to 4.0 deleted 468 of
them (~54%) and added 63 new lines (~7%). Reindexing it on my test
setup (Postgres 9.4.6 on Fedora 23, with default configurations) reduced
the size of the ix_vdc_options index from 80KB to 40KB (50%) and the
size of pk_vdc_options, the primary key's index, from 40KB to 32KB
(20%).

Change-Id: I2fff10b11bf3dafa71f96651ef2cecb491f89a86
Signed-off-by: Allon Mureinik <amureini@redhat.com>