core: No StorageType in VersionStorageFormatUtil
VersionStorageFormatUtil is supposed to map between a logical DC level
(e.g., 3.6) to a StorageFormatType constant (e.g. V3). The entire point
of having format versions is to transcend the differences of various
storage types.
DC level 3.1 failed to do so, and mixed V1 (for file based storage )and
V2 (for block storage) in the DC level. With their removal, the
interface can now be cleaned up, and simply map a version to a format.

Change-Id: Icd9250fc0ff88cc3776b8c5ef5621768e26bd2a3
Signed-off-by: Allon Mureinik <amureini@redhat.com>