core: Remove MoveVmCommand
oVirt 3.0 had a command to move a VM from one storage domain to another.
oVirt 3.1 introduced the notion of a VM having multiple disks on
multiple storage domains, with the ability to move them independently,
rendering this command meaningless, and it was indeed marked as
deprecated and preserved only in order to not break the REST API's
backwards compatibility. Change
I8953fe1c6c68e5bc95662342c8193fa77a1019d3 removed this broken API call,
and this patch finishes the job be removing the backend code that used
to support it

With the removal of MoveVMCommand,
MoveOrCopyTemplateCommand#getMoveOrCopyImageOperation() will always
return ImageOperation.Copy, so any code referencing ImageOperation.Move
in this hierarchy was also removed.

Change-Id: I2a738a78d9f5f7f3812e3f8443c9a49c3d012daa
Bug-Url: https://bugzilla.redhat.com/1320515
Signed-off-by: Allon Mureinik <amureini@redhat.com>