core: Remove VdcActionType#MoveOrCopyVmTemplate
oVirt 3.0 had a command to copy a VM Template from one storage domain
to another.
oVirt 3.1 introduced the notion of a VM Template having multiple disks
on multiple storage domains, with the ability to copy them independently,
rendering this command meaningless, and it is indeed not used in the
codebase, and hasn't been for a while.

This patch removes the constant representing it and all the tooling
around it (ExecutionMessages constant, LocalizedEnums translations,
etc.). Subsequent patches will remove the actual code for the
MoveOrCopyTemplateCommand, once the classes that extend it will be
cleaned up.

Change-Id: I64ae7eba3834e5b3347d740c81763d6308ccdc86
Signed-off-by: Allon Mureinik <amureini@redhat.com>