restapi: Restore incorrect behaviour of V3 remove
In version 3 of the API some DELETE operations have two signatures: with
and without a body. This is bad practice, and has been removed in
version 4 of the API. But we need to preserve them for backwards
compatibility, and the only way to differentiate one the other is to
use different "@Consumes" annotations. The signature that doesn't accept
a body should not have this annotation, so that it accepts requests that
don't include the "Accept" header, or "Accept: */*". The other signature
needs to have the annotation, specifying that it accepts XML and JSON,
so that it is selected when a body is included in the request. This
incorrect behaviour existed in version 3 of the API, but was lost when
the 3 to 4 transformation layer was introduced. This patch restores it.

Change-Id: I31f98d1a28c267b4fa48bea59ded539f789d76ef
Bug-Url: https://bugzilla.redhat.com/1318666
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>