core: FeatureSupported getOrDefault usage
FeatureSupported is in the common module, and as such, it should only
use JDK 7 methods (since GWT does not yet support JDK 8 methods, as
verified by animal-sniffer).

This patch removes the usage of getOrDefault and implements an
equivalent logic (under the assumption that the config map may not
contain null values) in-place by checking if the value returned from
the get method was null or not.

Change-Id: Ibb7775796005d8a65f1741e01458589e95f33ac1
Signed-off-by: Allon Mureinik <amureini@redhat.com>