core: Make MoveOrCopyTemplateCommand abstract
As a step to remove this unused (and useless) command, this patch
makes it abstract and removes the methods that are overriden by all the
classes that extend it.

The test was also removed (as this class can no longer be instantiated),
as well as audit log constants that are no longer used anywhere in the
codebase one the respective methods this patch cleans up were removed.

Change-Id: Id6267cb8f13e0f395b8737862dbe86768f4941ab
Signed-off-by: Allon Mureinik <amureini@redhat.com>