core: Fail fast ImportVmTemplateCommand#validate()
Cleaned up the ImportVmTemaplateCommand#validate() method to use the
early-return (fail-fast) idiom in order to clean up the code and make it
more readable, instead of dragging the retVal variable throughout the
method.

As part of this refactoring, the setActionMessageParameters() method was
overridden so that validate() doesn't have to handle adding the default
variables. Moreover, since it previously was NOT overridden, it
inherited its implementation from its superclass,
MoveOrCopyTemplateCommand, which caused an undefined behavior.

Change-Id: I335ada3c12b83db2d4a904fdfc30affec4d835c9
Signed-off-by: Allon Mureinik <amureini@redhat.com>