core: Inline MoveOrCopyTemplateCommand#isImagesAlreadyOnTarget
MoveOrCopyTemplateCommand#isImagesAlreadyOnTarget does nothing but
delegate to a method of the same name in the parameter class. It does
nothing to improve the code's readability, and in fact hinders it, as
the reader may suspect this method actually performs some check.

This patch removes the aforementioned method in order to reduce the
code's complexity and improve its readability.

Change-Id: Ic580c5c28f85ca91ccd2f7d47b9b57b5156e058f
Signed-off-by: Allon Mureinik <amureini@redhat.com>