HE: Make the unmanaged hosted engine error a bit more generic
This updates the ACTION_TYPE_FAILED_UNMANAGED_HOSTED_ENGINE
error message to include the full action and type to make
it ready for more actions than just VM activation.

The message is changed from:

Cannot ${action} unmanaged hosted engine VM. Please upgrade the cluster
level to 3.6 and wait for the hosted engine storage domain to be
properly imported.

To:

Cannot ${action} ${type}. You are using an unmanaged hosted engine VM.
Please upgrade the cluster level to 3.6 and wait for the hosted engine
storage domain to be properly imported.

Change-Id: Ie3230c4b7a783d200df0da18cf8462a21aa0b54f
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1319769
Signed-off-by: Martin Sivak <msivak@redhat.com>