core: prevent a race on add template version
When adding a template version, we add the quartz job and then mapping
the ID of the base template to that job so we will be able to stop
on-going updates of VMs to the latest version of the template.

The problem is that we first "register" the quartz job and only then
add it to the map (because we need to ID of the produced job). But
when the job is being executed before the mapping is set, the VM
updates will not take place.

This patch suggests a simple fix - to add the template with null job
before registering the quartz job, and update the mapping after the
job is created.

We should rewrite the mechanism that cancels on-going updates
someday.

Change-Id: I955a655f2eaef8510e3884e7895b9ca6cd0a6349
Bug-Url: https://bugzilla.redhat.com/1294511
Signed-off-by: Arik Hadas <ahadas@redhat.com>