core: Remove VmTemplateCommand#executeCommand
Throwing a NotImplementedException isn't a useful behavior - its just a
stub implementation intended to mark that whoever extends this class
should override this method.

A much better way to do this is to declare the class abstract (as it
already is), and leave this method unimplemented, so overriding it is
checked in compile time.

Change-Id: Ia96bf6ddfd5ca194a70f86742be609ec8a400d4e
Signed-off-by: Allon Mureinik <amureini@redhat.com>