core: Migration max bandwidth
Computation of max bandwidth for migration based on Cluster properties
'migrationBandwidthLimitType', 'customMigrationNetworkBandwidth' and
MigrationPolicy#maxMigrations.

If limit type == AUTO and no QoS is set on migration network than the
bandwidth is estimated as minimum of link speeds of source and
destination NICs connected to the migration network.

In case of limit type == AUTO && cross-cluster migration, the bandwidth
is estimate as a minimum of estimations for each cluster.

Change-Id: I7840b44ab114ff7c8276287fe1b6c19864d3b39e
Bug-Url: https://bugzilla.redhat.com/1252426
Feature-Page: http://www.ovirt.org/Features/Migration_Enhancements
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>