restapi: Use "fromValue" in V3 adapters
The recent introduction of type safe enums in version 4 of the API
introduced a regression in the V3 adapters. These adapters convert the
strings used in V3 to the enums used in V4, but some of them do so using
the "valueOf" method of the enum types, wich expects strings exactly
equal to the enum values, and these values are all upper case. The API,
both V3 and V4, should ignore case of the strings, which is achieved
using the "fromValue" method. This changes all the V3 adapters to use
the "fromValue" method.

Change-Id: I0637285b60dc7030dcbfa65d97927978e05e839d
Bug-Url: https://bugzilla.redhat.com/1320964
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>