restapi: Fix V3 VM graphics console server
The V3 VM graphics console server doesn't support the "proxyticket"
operation because it implements the base graphics console interface
used by virtual machines and templates, and this interface doesn't
contain the "proxyticket" operation. This patch changes the server so
that it implements the interface specialized for virtual machines, which
does contain the "proxyticket" operation.

Change-Id: I192b32f7822250147003af7b93db4e01b162d42a
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>