webadmin: SetupNetworkModel refactoring
The commit includes:
1. 'allNics' is a list of the existing host nics.
Until now it was updated upon each operation that was done in the dialog.
Now, once the nic list is queried it will stay untouched.
2. 'DataFromHostSetupNetworksModel hostSetupNetworksParametersData'
contains the data to build the parameters for the setup networks engine
command.
 2.1 It contains the 'whole' network config and not just the diffs
(the engine will ignore the unchanged values).
 2.2 'hostSetupNetworksParametersData' is updated upon each operation in
the dialog.
 2.3  The ui models (LogicalNetworkModel, NetworkInterfaceModel,
BondNetworkInterfaceModel, NetworkLabelModel) of the dialog are built using
the data in 'hostSetupNetworksParametersData'.
 2.3 The only proccess that updates and builds the ui models is the redraw/
initAllModels (same as before).
It means that the operations (NetworkOperation) update only the
'hostSetupNetworksParametersData'. After the operation is done update the
params, redraw should be called to recreated the ui models. (Same as it
was prior this commit).

The commit doesn't include:
1. Fixing the setup networks tooltips.

Change-Id: Ic9ab17fce26b36f8d9908d726fb9774eec31600a
Signed-off-by: Alona Kaplan <alkaplan@redhat.com>